mbt connection changes and recording changes
