+++ DepthCamera/DepthCameraApp.swift
@@ -10,13 +10,15 @@ import ConnectIQ
 
 @main
 struct DepthCameraApp: App {
-    @StateObject private var deviceViewModel = DeviceViewModel()
+    @StateObject private var deviceViewModel = DeviceViewModel() // this needs to be here to parse the list of devices availabe from the user's garmin connect account
+    @StateObject private var mbtViewModel = MBTViewModel() // this needs to be here to parse the list of devices availabe from the user's garmin connect account
 
     var body: some Scene {
         WindowGroup {
-            ContentView(deviceViewModel: deviceViewModel).onAppear {
+            ContentView(deviceViewModel: deviceViewModel, mbtViewModel: mbtViewModel).onAppear {
                 UIApplication.shared.isIdleTimerDisabled = true // Prevent screen locking
                 initializeGarminConnectIQ()
+                UIScreen.main.brightness = CGFloat(0.01)
             }
             .onDisappear {
                 UIApplication.shared.isIdleTimerDisabled = false // Re-enable screen locking
@@ -24,6 +26,7 @@ struct DepthCameraApp: App {
             .onOpenURL { url in
                 cdebug("Received URL: \(url)")
                 if let query = url.query(), query.contains("ciqBundle") {
+                    // need to store the list of availabe devices
                     deviceViewModel.parseDevices(from: url)
                 }
             }
