+++ DepthCamera/Views/MBTView.swift
@@ -6,71 +6,21 @@
 //
 import SwiftUI
 
-struct ThumbnailView: View {
-    private let thumbnailFrameWidth: CGFloat = 60.0
-    private let thumbnailFrameHeight: CGFloat = 60.0
-    private let thumbnailFrameCornerRadius: CGFloat = 10.0
-    private let thumbnailStrokeWidth: CGFloat = 2
-    
-    @ObservedObject var model: ARViewModel
+struct MBTView: View {
+    @ObservedObject var model : MBTViewModel
     @Environment(\.presentationMode) var presentationMode
-    
-    @State private var isShowingFilePicker = false
+    @State private var isShowingSync = false
     
     var body: some View {
         Button(action: {
-            isShowingFilePicker = true
+            isShowingSync = true
         }) {
-            Group {
-                Image(systemName: "photo.on.rectangle")
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .padding(16)
-                    .frame(width: thumbnailFrameWidth, height: thumbnailFrameHeight)
-                    .foregroundColor(.primary)
-                    .overlay(
-                        RoundedRectangle(cornerRadius: thumbnailFrameCornerRadius)
-                            .stroke(Color.white, lineWidth: thumbnailStrokeWidth)
-                    )
-            }
-            .onAppear {
-            }
+            Image("m3")
         }
-        .sheet(isPresented: $isShowingFilePicker) {
-            DocumentPicker(directoryURL: getDocumentsDirectory())
+        .sheet(isPresented: $isShowingSync) {
+            Text("MBT Connection Status")
         }
     }
     
-    func getDocumentsDirectory() -> URL {
-        let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
-        return paths[0]
-    }
 }
 
-struct ThumbnailImageView: View {
-    var uiImage: UIImage
-    var thumbnailFrameWidth: CGFloat
-    var thumbnailFrameHeight: CGFloat
-    var thumbnailFrameCornerRadius: CGFloat
-    var thumbnailStrokeWidth: CGFloat
-    
-    init(uiImage: UIImage, width: CGFloat, height: CGFloat, cornerRadius: CGFloat,
-         strokeWidth: CGFloat) {
-        self.uiImage = uiImage
-        self.thumbnailFrameWidth = width
-        self.thumbnailFrameHeight = height
-        self.thumbnailFrameCornerRadius = cornerRadius
-        self.thumbnailStrokeWidth = strokeWidth
-    }
-    var body: some View {
-        Image(uiImage: uiImage)
-            .resizable()
-            .aspectRatio(contentMode: .fill)
-            .frame(width: thumbnailFrameWidth, height: thumbnailFrameHeight)
-            .cornerRadius(thumbnailFrameCornerRadius)
-            .clipped()
-            .overlay(RoundedRectangle(cornerRadius: thumbnailFrameCornerRadius)
-                .stroke(Color.primary, lineWidth: thumbnailStrokeWidth))
-            .shadow(radius: 10)
-    }
-}
