+++ DepthCamera/Views/ThumbnailView.swift
@@ -1,3 +1,11 @@
+//
+//  ThumbnailView.swift
+//  DepthCamera
+//
+//  Created by Brian Toone on 12/10/24.
+//
+import SwiftUI
+
 struct ThumbnailView: View {
     private let thumbnailFrameWidth: CGFloat = 60.0
     private let thumbnailFrameHeight: CGFloat = 60.0
@@ -8,33 +16,61 @@ struct ThumbnailView: View {
     @Environment(\.presentationMode) var presentationMode
     
     @State private var isShowingFilePicker = false
-
+    
     var body: some View {
-          Button(action: {
-              isShowingFilePicker = true
-          }) {
-              Group {
-                  Image(systemName: "photo.on.rectangle")
-                      .resizable()
-                      .aspectRatio(contentMode: .fit)
-                      .padding(16)
-                      .frame(width: thumbnailFrameWidth, height: thumbnailFrameHeight)
-                      .foregroundColor(.primary)
-                      .overlay(
-                          RoundedRectangle(cornerRadius: thumbnailFrameCornerRadius)
-                              .stroke(Color.white, lineWidth: thumbnailStrokeWidth)
-                      )
-              }
-              .onAppear {
-              }
-          }
-          .sheet(isPresented: $isShowingFilePicker) {
-              DocumentPicker(directoryURL: getDocumentsDirectory())
-          }
-      }
-      
-      func getDocumentsDirectory() -> URL {
-          let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
-          return paths[0]
-      }
-  }
+        Button(action: {
+            isShowingFilePicker = true
+        }) {
+            Group {
+                Image(systemName: "photo.on.rectangle")
+                    .resizable()
+                    .aspectRatio(contentMode: .fit)
+                    .padding(16)
+                    .frame(width: thumbnailFrameWidth, height: thumbnailFrameHeight)
+                    .foregroundColor(.primary)
+                    .overlay(
+                        RoundedRectangle(cornerRadius: thumbnailFrameCornerRadius)
+                            .stroke(Color.white, lineWidth: thumbnailStrokeWidth)
+                    )
+            }
+            .onAppear {
+            }
+        }
+        .sheet(isPresented: $isShowingFilePicker) {
+            DocumentPicker(directoryURL: getDocumentsDirectory())
+        }
+    }
+    
+    func getDocumentsDirectory() -> URL {
+        let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
+        return paths[0]
+    }
+}
+
+struct ThumbnailImageView: View {
+    var uiImage: UIImage
+    var thumbnailFrameWidth: CGFloat
+    var thumbnailFrameHeight: CGFloat
+    var thumbnailFrameCornerRadius: CGFloat
+    var thumbnailStrokeWidth: CGFloat
+    
+    init(uiImage: UIImage, width: CGFloat, height: CGFloat, cornerRadius: CGFloat,
+         strokeWidth: CGFloat) {
+        self.uiImage = uiImage
+        self.thumbnailFrameWidth = width
+        self.thumbnailFrameHeight = height
+        self.thumbnailFrameCornerRadius = cornerRadius
+        self.thumbnailStrokeWidth = strokeWidth
+    }
+    var body: some View {
+        Image(uiImage: uiImage)
+            .resizable()
+            .aspectRatio(contentMode: .fill)
+            .frame(width: thumbnailFrameWidth, height: thumbnailFrameHeight)
+            .cornerRadius(thumbnailFrameCornerRadius)
+            .clipped()
+            .overlay(RoundedRectangle(cornerRadius: thumbnailFrameCornerRadius)
+                .stroke(Color.primary, lineWidth: thumbnailStrokeWidth))
+            .shadow(radius: 10)
+    }
+}
